[backend/feat] 친구 리스트 필드 추가
